+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add worker tasks queue options to pool options.
+
 ## [2.4.6] - 2023-04-10
 
 ### Fixed